+++ mayan/apps/dynamic_search/tasks.py
@@ -9,7 +9,8 @@ from mayan.celery import app
 from .classes import SearchBackend, SearchModel
 from .exceptions import DynamicSearchException, DynamicSearchRetry
 from .literals import (
-    TASK_DEINDEX_INSTANCE_MAX_RETRIES, TASK_INDEX_INSTANCE_MAX_RETRIES
+    TASK_DEINDEX_INSTANCE_MAX_RETRIES, TASK_INDEX_INSTANCE_MAX_RETRIES,
+    TASK_INDEX_SEARCH_MODEL_MAX_RETRIES
 )
 from .settings import setting_indexing_chunk_size
 
@@ -35,7 +36,8 @@ def task_deindex_instance(self, app_label, model_name, object_id):
 
 
 @app.task(
-    bind=True, ignore_result=True, max_retries=None, retry_backoff=True
+    bind=True, ignore_result=True,
+    max_retries=TASK_INDEX_SEARCH_MODEL_MAX_RETRIES, retry_backoff=True
 )
 def task_index_search_model(self, search_model_full_name, range_string=None):
     search_model = SearchModel.get(name=search_model_full_name)
@@ -46,11 +48,16 @@ def task_index_search_model(self, search_model_full_name, range_string=None):
 
     try:
         SearchBackend.get_instance().index_search_model(**kwargs)
+    except (DynamicSearchRetry, LockError) as exception:
+        raise self.retry(exc=exception)
     except Exception as exception:
-        raise DynamicSearchException(
-            'Unexpected error calling `index_search_model` with '
-            'keyword arguments {}.'.format(kwargs)
-        ) from exception
+        error_message = (
+            'Unexpected error calling `task_index_search_model` with '
+            'keyword arguments {}.'
+        ).format(kwargs)
+
+        logger.error(error_message)
+        raise DynamicSearchException(error_message) from exception
 
 
 @app.task(
@@ -80,8 +87,24 @@ def task_index_instance(
         )
     except (DynamicSearchRetry, LockError) as exception:
         raise self.retry(exc=exception)
+    except Exception as exception:
+        kwargs = {
+            'app_label': app_label,
+            'model_name': model_name,
+            'object_id': object_id,
+            'exclude_app_label': exclude_app_label,
+            'exclude_model_name': exclude_model_name,
+            'exclude_kwargs': exclude_kwargs
+        }
+        error_message = (
+            'Unexpected error calling `task_index_instance` with keyword '
+            'arguments {}.'
+        ).format(kwargs)
 
-    logger.info('Finished')
+        logger.error(error_message)
+        raise DynamicSearchException(error_message) from exception
+    else:
+        logger.info('Finished')
 
 
 @app.task(ignore_result=True)