+++ mayan/apps/dynamic_search/tests/mixins.py
@@ -1,9 +1,25 @@
+from mayan.apps.databases.literals import DATABASE_MINIMUM_ID
 from mayan.apps.documents.search import document_search
 
 from ..classes import SearchBackend, SearchModel
 from ..literals import QUERY_PARAMETER_ANY_FIELD, SEARCH_MODEL_NAME_KWARG
+from ..tasks import task_reindex_backend, task_index_search_model
 
 from .backends import TestSearchBackend
+
+
+class SearchTaskTestMixin:
+    def _execute_task_reindex_backend(self):
+        task_reindex_backend.apply_async().get()
+
+    def _execute_task_index_search_model(self):
+        task_index_search_model.apply_async(
+            kwargs={
+                'range_string': '{}-{}'.format(
+                    DATABASE_MINIMUM_ID, self.test_object.pk
+                ), 'search_model_full_name': self.test_model_search.get_full_name()
+            }
+        ).get()
 
 
 class SearchTestMixin: